Better Testing

+++ core/util.cpp
@@ -4,7 +4,7 @@
 void Util::syncGamesFast(const Game & from, Game & to)
 {
 	auto const & history = from.getMoveHistory();
-	for (uint i = to.getMoveHistory().size(); i < history.size(); ++i)
+	for (size_t i = to.getMoveHistory().size(); i < history.size(); ++i)
 		to.simStep(history[i]);
 }
 
@@ -17,7 +17,7 @@ void Util::syncGames(const Game & from, Game & to)
 		to.undo();
 
 	int offset = toHistory.size();
-	for (int i = 0; i < toHistory.size(); ++i)
+	for (size_t i = 0; i < toHistory.size(); ++i)
 	{
 		if (fromHistory[i] != toHistory[i])
 		{
@@ -25,7 +25,7 @@ void Util::syncGames(const Game & from, Game & to)
 			break;
 		}
 	}
-	for (int i = offset; i < toHistory.size(); ++i)
+	for (size_t i = offset; i < toHistory.size(); ++i)
 		to.undo();
 
 	syncGamesFast(from, to);
